Split item form projection helpers
